Use smaller integers for day 4

+++ src/p4.rs
@@ -2,35 +2,35 @@ use super::*;
 
 #[derive(Debug)]
 struct Board {
-    grid: [u32; 25],
+    grid: [u8; 25],
     set: IntMap,
-    row_count: [u32; Self::LENGTH],
-    col_count: [u32; Self::LENGTH],
+    row_count: [u8; Self::LENGTH],
+    col_count: [u8; Self::LENGTH],
     solved: bool,
 }
 
 #[derive(Debug)]
 struct IntMap {
-    data: [usize; Self::MAX_VAL],
+    data: [u8; Self::MAX_VAL],
 }
 
 impl IntMap {
     const MAX_VAL: usize = 100;
-    const EMPTY: usize = 255;
+    const EMPTY: u8 = 255;
     fn new() -> Self {
         Self {
             data: [Self::EMPTY; Self::MAX_VAL],
         }
     }
-    fn insert(&mut self, index: u32, val: usize) {
-        self.data[index as usize] = val;
+    fn insert(&mut self, index: usize, val: u8) {
+        self.data[index] = val;
     }
-    fn remove(&mut self, index: u32) -> Option<usize> {
+    fn remove(&mut self, index: usize) -> Option<u8> {
         let val = self.data[index as usize];
         if val == Self::EMPTY {
             None
         } else {
-            self.data[index as usize] = Self::EMPTY;
+            self.data[index] = Self::EMPTY;
             Some(val)
         }
     }
@@ -38,7 +38,7 @@ impl IntMap {
 
 impl Board {
     const LENGTH: usize = 5;
-    fn new(grid: [u32; 25], set: IntMap) -> Self {
+    fn new(grid: [u8; 25], set: IntMap) -> Self {
         Self {
             grid,
             set,
@@ -47,11 +47,11 @@ impl Board {
             solved: false,
         }
     }
-    fn update(&mut self, draw: u32) -> bool {
+    fn update(&mut self, draw: u8) -> bool {
         if self.solved {
             return false;
         }
-        if let Some(idx) = self.set.remove(draw) {
+        if let Some(idx) = self.set.remove(draw as usize) {
             let idx = idx as usize;
             self.row_count[idx / Self::LENGTH] += 1;
             self.col_count[idx % Self::LENGTH] += 1;
@@ -89,21 +89,20 @@ pub fn solve(input: &[u8]) -> Solution<u32, u32> {
                 num = 0;
             }
             _ => {
-                num = num * 10 + (input[idx] - b'0') as u32;
+                num = num * 10 + (input[idx] - b'0');
             }
         }
         idx += 1;
     }
 
-    let mut grid: [u32; 25] = [0; 25];
+    let mut grid: [u8; 25] = [0; 25];
     let mut set = IntMap::new();
     let mut boards = Vec::new();
     let mut grid_index = 0;
     while idx < input.len() {
-        let num =
-            (input[idx].saturating_sub(b'0') * 10 + input[idx + 1].saturating_sub(b'0')) as u32;
+        let num = input[idx].saturating_sub(b'0') * 10 + input[idx + 1].saturating_sub(b'0');
         grid[grid_index] = num;
-        set.insert(num, grid_index);
+        set.insert(num as usize, grid_index as u8);
         grid_index += 1;
         idx += 3;
         if grid_index >= 25 {
@@ -118,7 +117,7 @@ pub fn solve(input: &[u8]) -> Solution<u32, u32> {
     Solution::new(part1, part2)
 }
 
-fn run_until_solved(boards: &mut Vec<Board>, order: &[u32]) -> (u32, u32) {
+fn run_until_solved(boards: &mut Vec<Board>, order: &[u8]) -> (u32, u32) {
     let mut part1 = 0;
     let mut solved = 0;
     let len = boards.len();
@@ -127,9 +126,9 @@ fn run_until_solved(boards: &mut Vec<Board>, order: &[u32]) -> (u32, u32) {
             if board.update(num) {
                 solved += 1;
                 if part1 == 0 {
-                    part1 = num * board.sum_unmarked() as u32;
+                    part1 = num as u32 * board.sum_unmarked() as u32;
                 } else if solved == len {
-                    return (part1, num * board.sum_unmarked() as u32);
+                    return (part1, num as u32 * board.sum_unmarked() as u32);
                 }
             }
         }
